HDR 背景をぼかせるようにする
素材を見比べるとき、背景の細部が目移りの原因になっていた。
ライティングと露出 > 背景をぼかす を追加する。

ぼかしパスは足していない。IBL 用にすでに作ってあるプリフィルタ済みキューブ
（ラフネス別に GGX で畳み込んである）の粗いミップを引く。ミップを落とすだけの
環境キューブより滑らかで、追加のコストもない。

- Skybox の mipLevel を uint から float にした。ぼかしの強さを段の間で
  決められるようにするため
- 背景の描画で引くテクスチャとミップを変えるだけなので、IBL の寄与と陰影は
  変わらない
- プロジェクトに skyboxBlur として保存する

強さはミップ 0.6 相当。1.0 以上にすると背景がほぼ単色になり、どこにいるのか
分からなくなる。草地・道・木立の配置は読めるが細部は目に留まらない強さにした。

検証: x64-debug ビルド、テスト成功、DXC で Skybox の VS / PS を単体コンパイル。
tree_lined_driveway_4k.hdr でミップ 0.0 / 0.6 / 1.0 / 2.0 / 3.0 を撮り比べ、
0.6 を採用した。--screenshot-ui でチェックボックスの並びも目視。
D3D12 検証エラーと ImGui の警告 0。

+++ shaders/Skybox.hlsl
@@ -12,7 +12,8 @@ struct SkyboxConstants
     float intensity;
 
     uint environmentIndex;
-    uint mipLevel;
+    // 引くミップ。小数を許すのは、ぼかしの強さを段の間で決められるようにするため。
+    float mipLevel;
     float2 pad0;
 };
 
@@ -47,6 +48,6 @@ float4 PsMain(VsOutput input) : SV_Target
     TextureCube<float4> environment = ResourceDescriptorHeap[g_skybox.environmentIndex];
     const float3 direction = normalize(input.direction);
     const float3 radiance =
-        environment.SampleLevel(g_samplerLinearClamp, direction, float(g_skybox.mipLevel)).rgb;
+        environment.SampleLevel(g_samplerLinearClamp, direction, g_skybox.mipLevel).rgb;
     return float4(radiance * g_skybox.intensity, 1.0f);
 }
